# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/payment-rules": {
      "post": {
        "description": "Creates a new payment rule that defines payment rates for a specific role and shift type combination",
        "operationId": "PaymentRulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePaymentRuleDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The payment rule has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRuleResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new payment rule",
        "tags": ["Payment Rules"]
      },
      "get": {
        "description": "Retrieves all payment rules with optional filtering",
        "operationId": "PaymentRulesController_findAll",
        "parameters": [
          {
            "name": "shiftTypeId",
            "required": false,
            "in": "query",
            "description": "Filter by shift type ID",
            "schema": {
              "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "description": "Filter by role ID",
            "schema": {
              "example": "3fa85f64-5717-4562-b3fc-2c963f66afa7",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "description": "Filter by organization ID",
            "schema": {
              "example": "3fa85f64-5717-4562-b3fc-2c963f66afa8",
              "type": "string"
            }
          },
          {
            "name": "paymentType",
            "required": false,
            "in": "query",
            "description": "Filter by payment type",
            "schema": {
              "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
              "type": "string"
            }
          },
          {
            "name": "effectiveDate",
            "required": false,
            "in": "query",
            "description": "Filter by effective date - returns rules effective at the specified date",
            "schema": {
              "format": "date-time",
              "example": "2025-06-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter active rules only (rules with no end date or end date in the future)",
            "schema": { "example": true, "type": "boolean" }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip for pagination",
            "schema": { "minimum": 0, "example": 0, "type": "number" }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take for pagination",
            "schema": { "minimum": 1, "example": 10, "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all payment rules that match the query.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentRuleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find all payment rules",
        "tags": ["Payment Rules"]
      }
    },
    "/api/payment-rules/{id}": {
      "get": {
        "description": "Retrieves a specific payment rule by its unique identifier",
        "operationId": "PaymentRulesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The payment rule ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the payment rule with the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRuleResponseDto"
                }
              }
            }
          },
          "404": { "description": "Payment rule not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find one payment rule by ID",
        "tags": ["Payment Rules"]
      },
      "patch": {
        "description": "Updates an existing payment rule with the provided data",
        "operationId": "PaymentRulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The payment rule ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePaymentRuleDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The payment rule has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRuleResponseDto"
                }
              }
            }
          },
          "404": { "description": "Payment rule not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update a payment rule",
        "tags": ["Payment Rules"]
      },
      "delete": {
        "description": "Removes a payment rule from the system",
        "operationId": "PaymentRulesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The payment rule ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The payment rule has been successfully deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRuleResponseDto"
                }
              }
            }
          },
          "404": { "description": "Payment rule not found." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Delete a payment rule",
        "tags": ["Payment Rules"]
      }
    },
    "/api/payment-rules/organization/{organizationId}": {
      "get": {
        "description": "Retrieves all payment rules for a specific organization",
        "operationId": "PaymentRulesController_findByOrganization",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The organization ID",
            "schema": { "type": "string" }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": { "type": "boolean" }
          },
          {
            "name": "effectiveDate",
            "required": false,
            "in": "query",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "paymentType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "shiftTypeId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all payment rules for the given organization.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentRuleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find payment rules by organization",
        "tags": ["Payment Rules"]
      }
    },
    "/api/payment-rules/role/{roleId}": {
      "get": {
        "description": "Retrieves all payment rules for a specific role",
        "operationId": "PaymentRulesController_findByRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "The role ID",
            "schema": { "type": "string" }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": { "type": "boolean" }
          },
          {
            "name": "effectiveDate",
            "required": false,
            "in": "query",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "paymentType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "shiftTypeId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all payment rules for the given role.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentRuleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find payment rules by role",
        "tags": ["Payment Rules"]
      }
    },
    "/api/payment-rules/shift-type/{shiftTypeId}": {
      "get": {
        "description": "Retrieves all payment rules for a specific shift type",
        "operationId": "PaymentRulesController_findByShiftType",
        "parameters": [
          {
            "name": "shiftTypeId",
            "required": true,
            "in": "path",
            "description": "The shift type ID",
            "schema": { "type": "string" }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": { "type": "boolean" }
          },
          {
            "name": "effectiveDate",
            "required": false,
            "in": "query",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "paymentType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all payment rules for the given shift type.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentRuleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Find payment rules by shift type",
        "tags": ["Payment Rules"]
      }
    },
...
  },
  "info": {
    "title": "Multi-Sector Management API",
    "description": "API for managing organizations across multiple sectors",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "CreatePaymentRuleDto": {
        "type": "object",
        "properties": {
          "shiftTypeId": {
            "type": "string",
            "description": "The ID of the shift type for this payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "roleId": {
            "type": "string",
            "description": "The ID of the role for this payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa7"
          },
          "paymentType": {
            "type": "string",
            "description": "The type of payment (hourly, weekly, monthly, or per shift)",
            "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
            "example": "HOURLY"
          },
          "baseRate": {
            "type": "number",
            "description": "The base payment rate",
            "example": 25.5,
            "minimum": 0
          },
          "specialtyBonus": {
            "type": "number",
            "description": "Additional bonus for specialized skills (optional)",
            "example": 5,
            "minimum": 0
          },
          "experienceMultiplier": {
            "type": "number",
            "description": "Multiplier applied based on experience level (optional)",
            "example": 1.2,
            "minimum": 1
          },
          "effectiveDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this payment rule becomes effective",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this payment rule expires (optional)",
            "example": "2025-12-31T00:00:00.000Z"
          },
          "organizationId": {
            "type": "string",
            "description": "The ID of the organization this payment rule belongs to",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa8"
          }
        },
        "required": [
          "shiftTypeId",
          "roleId",
          "paymentType",
          "baseRate",
          "effectiveDate",
          "organizationId"
        ]
      },
      "PaymentRuleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa9"
          },
          "shiftTypeId": {
            "type": "string",
            "description": "The ID of the shift type for this payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "roleId": {
            "type": "string",
            "description": "The ID of the role for this payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa7"
          },
          "paymentType": {
            "type": "string",
            "description": "The type of payment",
            "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
            "example": "HOURLY"
          },
          "baseRate": {
            "type": "number",
            "description": "The base payment rate",
            "example": 25.5
          },
          "specialtyBonus": {
            "type": "number",
            "description": "Additional bonus for specialized skills",
            "example": 5
          },
          "experienceMultiplier": {
            "type": "number",
            "description": "Multiplier applied based on experience level",
            "example": 1.2
          },
          "effectiveDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this payment rule becomes effective",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this payment rule expires",
            "example": "2025-12-31T00:00:00.000Z",
            "nullable": true
          },
          "organizationId": {
            "type": "string",
            "description": "The ID of the organization this payment rule belongs to",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa8"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the payment rule was created",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the payment rule was last updated",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "shiftType": {
            "type": "object",
            "description": "The related shift type details"
          },
          "role": {
            "type": "object",
            "description": "The related role details"
          }
        },
        "required": [
          "id",
          "shiftTypeId",
          "roleId",
          "paymentType",
          "baseRate",
          "specialtyBonus",
          "experienceMultiplier",
          "effectiveDate",
          "organizationId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePaymentRuleDto": {
        "type": "object",
        "properties": {
          "shiftTypeId": {
            "type": "string",
            "description": "The ID of the shift type for this payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "roleId": {
            "type": "string",
            "description": "The ID of the role for this payment rule",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa7"
          },
          "paymentType": {
            "type": "string",
            "description": "The type of payment (hourly, weekly, monthly, or per shift)",
            "enum": ["HOURLY", "WEEKLY", "MONTHLY", "PER_SHIFT"],
            "example": "HOURLY"
          },
          "baseRate": {
            "type": "number",
            "description": "The base payment rate",
            "example": 25.5,
            "minimum": 0
          },
          "specialtyBonus": {
            "type": "number",
            "description": "Additional bonus for specialized skills",
            "example": 5,
            "minimum": 0
          },
          "experienceMultiplier": {
            "type": "number",
            "description": "Multiplier applied based on experience level",
            "example": 1.2,
            "minimum": 1
          },
          "effectiveDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this payment rule becomes effective",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date when this payment rule expires",
            "example": "2025-12-31T00:00:00.000Z",
            "nullable": true
          }
        }
      },
...
    }
  }
}